Split loki libs
